refactor: #649: move donor import file schema models to new project.
