+++ Atlas.DonorImport.FileSchema.Models/CHANGELOG_DonorImportFileSchema.md
@@ -1,0 +1,13 @@
+﻿# Changelog
+All notable data structure changes to this project will be documented in this file.
+
+This includes both schema and data workflow changes.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## Versions
+
+### 1.5.0
+* No outward change to schema definition.
+* Models moved from main Atlas.DonorImport project to new, standalone project, with versioning & changelog, to allow publishing as NuGet package.
