+++ Atlas.DonorImport.Test.Integration/IntegrationTests/DonorUpdates/DonorUpdatesPublisherTests.cs
@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 using Atlas.Common.ServiceBus;
 using Atlas.Common.Test.SharedTestHelpers;
 using Atlas.DonorImport.ExternalInterface.Models;
-using Atlas.DonorImport.Models.FileSchema;
+using Atlas.DonorImport.FileSchema.Models;
 using Atlas.DonorImport.Services;
 using Atlas.DonorImport.Test.Integration.TestHelpers;
 using Atlas.DonorImport.Test.TestHelpers.Builders.ExternalModels;
