+++ Atlas.DonorImport.Test.Integration/IntegrationTests/Import/UploadOrderTests.cs
@@ -1,11 +1,10 @@
 ﻿using System;
 using System.Threading.Tasks;
-using Atlas.Common.GeneticData;
 using Atlas.Common.Notifications;
 using Atlas.Common.Public.Models.GeneticData;
 using Atlas.Common.Test.SharedTestHelpers;
 using Atlas.DonorImport.ExternalInterface.Models;
-using Atlas.DonorImport.Models.FileSchema;
+using Atlas.DonorImport.FileSchema.Models;
 using Atlas.DonorImport.Services;
 using Atlas.DonorImport.Test.Integration.TestHelpers;
 using Atlas.DonorImport.Test.TestHelpers.Builders;
