+++ Atlas.DonorImport.Test/Services/DonorImportFileParserTests.cs
@@ -1,6 +1,6 @@
 using System.Linq;
 using Atlas.DonorImport.Exceptions;
-using Atlas.DonorImport.Models.FileSchema;
+using Atlas.DonorImport.FileSchema.Models;
 using Atlas.DonorImport.Services;
 using Atlas.DonorImport.Test.TestHelpers.Builders;
 using EnumStringValues;
