+++ Atlas.DonorImport.Test/Services/ImportedLocusInterpreterTests.cs
@@ -1,6 +1,6 @@
 using Atlas.Common.ApplicationInsights;
 using Atlas.Common.GeneticData.Hla.Services;
-using Atlas.DonorImport.Models.FileSchema;
+using Atlas.DonorImport.FileSchema.Models;
 using Atlas.DonorImport.Models.Mapping;
 using FluentAssertions;
 using NSubstitute;
