+++ Atlas.DonorImport.Test/TestHelpers/Builders/DonorImportFileContentsBuilder.cs
@@ -1,7 +1,6 @@
-using System;
 using System.Collections.Generic;
 using System.Linq;
-using Atlas.DonorImport.Models.FileSchema;
+using Atlas.DonorImport.FileSchema.Models;
 using Atlas.DonorImport.Test.TestHelpers.Models;
 using Atlas.DonorImport.Test.TestHelpers.Models.MalformedDonorFileModels;
 using LochNessBuilder;
