+++ Atlas.DonorImport.Test/TestHelpers/Builders/DonorUpdateBuilder.cs
@@ -1,7 +1,6 @@
-using Atlas.Common.GeneticData;
 using Atlas.Common.Public.Models.GeneticData;
 using Atlas.Common.Test.SharedTestHelpers;
-using Atlas.DonorImport.Models.FileSchema;
+using Atlas.DonorImport.FileSchema.Models;
 using Atlas.DonorImport.Test.TestHelpers.Models.MalformedDonorFileModels;
 using LochNessBuilder;
