+++ Atlas.DonorImport.Test/TestHelpers/Builders/ExternalModels/DonorImportFileBuilder.cs
@@ -1,7 +1,7 @@
 using System;
 using Atlas.Common.Test.SharedTestHelpers;
 using Atlas.DonorImport.ExternalInterface.Models;
-using Atlas.DonorImport.Models.FileSchema;
+using Atlas.DonorImport.FileSchema.Models;
 using Atlas.DonorImport.Test.TestHelpers.Models;
 using LochNessBuilder;
