+++ Atlas.DonorImport.Test/TestHelpers/Models/MalformedDonorFileModels/DonorFileWithDonorUpdateWithMissingField.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.IO;
 using System.Text;
-using Atlas.DonorImport.Models.FileSchema;
+using Atlas.DonorImport.FileSchema.Models;
 using Newtonsoft.Json;
 
 namespace Atlas.DonorImport.Test.TestHelpers.Models.MalformedDonorFileModels
