+++ Atlas.DonorImport.Test/TestHelpers/Models/MalformedDonorFileModels/DonorFileWithoutDonor.cs
@@ -1,6 +1,6 @@
 ﻿using System.IO;
 using System.Text;
-using Atlas.DonorImport.Models.FileSchema;
+using Atlas.DonorImport.FileSchema.Models;
 using Newtonsoft.Json;
 
 namespace Atlas.DonorImport.Test.TestHelpers.Models.MalformedDonorFileModels
