+++ Atlas.DonorImport/Models/Mapping/ImportedLocusInterpreter.cs
@@ -1,9 +1,8 @@
 using System.Collections.Generic;
 using Atlas.Common.ApplicationInsights;
 using Atlas.Common.GeneticData.Hla.Services;
-using Atlas.Common.GeneticData.PhenotypeInfo;
 using Atlas.Common.Public.Models.GeneticData.PhenotypeInfo;
-using Atlas.DonorImport.Models.FileSchema;
+using Atlas.DonorImport.FileSchema.Models;
 using MoreLinq;
 
 namespace Atlas.DonorImport.Models.Mapping
