+++ Atlas.DonorImport/Services/DonorRecordChangeApplier.cs
@@ -2,7 +2,6 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using Atlas.Common.GeneticData;
 using Atlas.Common.Notifications;
 using Atlas.Common.Public.Models.GeneticData;
 using Atlas.Common.Utils;
@@ -12,7 +11,8 @@ using Atlas.DonorImport.Data.Repositories;
 using Atlas.DonorImport.Exceptions;
 using Atlas.DonorImport.ExternalInterface.Models;
 using Atlas.DonorImport.ExternalInterface.Settings;
-using Atlas.DonorImport.Models.FileSchema;
+using Atlas.DonorImport.Models;
+using Atlas.DonorImport.FileSchema.Models;
 using Atlas.DonorImport.Models.Mapping;
 using Atlas.DonorImport.Services.DonorUpdates;
 using Atlas.MatchingAlgorithm.Client.Models.Donors;
